requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
argcomplete==3.1.4
async-timeout==5.0.1
attrs==23.2.0
Automat==22.10.0
Babel==2.10.3
base58==2.1.1
bcrypt==3.2.2
bitarray==3.3.1
blinker==1.7.0
certifi==2023.11.17
chardet==5.2.0
ckzg==2.1.1
click==8.1.6
cloud-init==24.4
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
constantly==23.10.4
cryptography==41.0.7
cytoolz==1.0.1
dbus-python==1.3.2
distro==1.9.0
distro-info==1.7+build1
eth-account==0.13.6
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.0
eth-utils==5.2.0
eth_abi==5.2.0
frozenlist==1.5.0
hexbytes==1.3.0
httplib2==0.20.4
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
ipfshttpclient==0.8.0a2
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multiaddr==0.0.9
multidict==6.2.0
netaddr==1.3.0
netifaces==0.11.0
oauthlib==3.2.2
packaging==23.2
parsimonious==0.10.0
pipx==1.4.3
platformdirs==4.2.0
propcache==0.3.1
psutil==5.9.8
py-solc-x==2.0.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.22.0
pycurl==7.45.3
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.17.2
PyGObject==3.48.2
PyHamcrest==2.1.0
PyJWT==2.7.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyrsistent==0.20.0
pyserial==3.5
python-apt==2.7.7+ubuntu3
python-dotenv==1.1.0
pytz==2024.1
pyunormalize==16.0.0
PyYAML==6.0.1
regex==2024.11.6
requests==2.31.0
rich==13.7.1
rlp==4.1.0
service-identity==24.1.0
six==1.16.0
systemd-python==235
toolz==1.0.0
Twisted==24.3.0
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.1
ubuntu-pro-client==8001
unattended-upgrades==0.1
urllib3==2.0.7
userpath==1.9.1
varint==1.0.2
wadllib==1.3.6
web3==7.10.0
websockets==15.0.1
yarl==1.19.0
zope.interface==6.1
